Passage des coordonnées en BigDecimal + Text en base

+++ src/main/kotlin/org/breizhcamp/rennes/tech/application/controllers/NextEventIcsCtrl.kt
@@ -76,7 +76,7 @@ class NextEventIcsCtrl(
             vevent = vevent.withProperty(Location(venue.completeAddress))
             if (venue.latitude != null && venue.longitude != null) {
                 vevent =
-                    vevent.withProperty(Geo(BigDecimal.valueOf(venue.latitude), BigDecimal.valueOf(venue.longitude)))
+                    vevent.withProperty(Geo(venue.latitude, venue.longitude))
             }
         }
