+++ src/main/kotlin/org/breizhcamp/rennes/tech/domain/entities/Venue.kt
@@ -1,5 +1,7 @@
 package org.breizhcamp.rennes.tech.domain.entities
 
+import java.math.BigDecimal
+
 sealed class Venue
 
 data class PhysicalVenue(
@@ -7,8 +9,8 @@ data class PhysicalVenue(
     val address: String,
     val city: String,
     val country: String,
-    val latitude: Double?,
-    val longitude: Double?,
+    val latitude: BigDecimal?,
+    val longitude: BigDecimal?,
 ) : Venue() {
     val completeAddress: String
         get() = listOf(name, address, city, country).filter { it.isNotBlank() }.joinToString(", ")
